+++ packages/hive-core/src/index.ts
@@ -12,4 +12,5 @@ export * from "./settings-schema.js";
 export * from "./settings.js";
 export * from "./machines.js";
 export * from "./ssh.js";
+export * from "./ipc.js";
 export * as templates from "./templates.js";